Renamed level helper module to level-io.js to prevent confusion in the future.
