+++ src/client/editor/project.js
@@ -1,6 +1,6 @@
 import { editorLog } from "./log";
 import * as Level from "../../level/level";
-import * as levelManager from "../../level/manager";
+import * as levelIO from "../../level/level-io";
 import SerializeWorker from "./serialize.worker";
 import { worldTab } from "./world";
 import { prefabsTab } from "./prefabs";
@@ -229,7 +229,7 @@ let projectTab = function() {
 		},
 
 		importProject: function(loadedFile) {
-			let loadedProject = levelManager.load(loadedFile);
+			let loadedProject = levelIO.load(loadedFile);
 
 			if(!loadedProject) {
 				editorLog("Unable to load project. Check console for details.", "error");
